refactor(view): extract design_import_shortcuts.cpp (A3 first cut)
design_import.cpp was 4670 lines mixing source parsing, classname
extraction, keyboard-shortcut extraction, Claude-bundle envelope,
react/v0/figma/stitch/rn/pencil parsers, and runtime-import wiring.
The 2026-05 spec called for splitting this into parse / bundle /
codegen / tokens TUs.

This first cut moves the keyboard-shortcut concern off the monolith:

* New TU `core/view/src/design_import_shortcuts.cpp` (~270 lines)
  owns `extract_keyboard_shortcuts`, `serialize_detected_shortcuts`,
  `key_string_to_keycode`, `modifier_strings_to_mask`, and the three
  file-local helpers (`line_for_offset`, `collect_modifiers`,
  `extract_handler_excerpt`).
* `design_import.cpp` drops from 4670 → 4417 lines (-253). A marker
  comment at the original section header points readers at the new TU.
* `core/view/CMakeLists.txt` adds the new source to `pulp-view`.
* Public API surface in `core/view/include/pulp/view/design_import.hpp`
  is unchanged. No caller touches.

The remaining A3 splits (claude_bundle.cpp for the bundle envelope +
parsers, design_codegen.cpp for IR → JS/native emission,
design_tokens.cpp for W3C/Figma/Stitch token import-export) are each
their own focused PR — bundling them all into one diff makes the
review window too narrow for the kind of careful boundary work each
needs. Tracked as the explicit A3 follow-up in
planning/2026-05-16-refactor-ops-high-leverage.md.

Validation:
* Existing tests in `test/test_design_import_shortcuts.cpp` /
  `test/test_design_import.cpp` continue to exercise the API
  through its public declarations — they don't care which TU the
  implementation lives in.
* CMake adds the new file to the same target so link succeeds.

Part of refactor-ops Batch 2.
